requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
langchain
openai
youtube-transcript-api
tiktoken